Refactor devices and zones into smaller files
Split large command files for better maintainability:

devices.go (185 lines) - types, list, get, values, findDevice helper
devices_control.go (113 lines) - set, on, off
devices_manage.go (62 lines) - rename, delete
devices_settings.go (108 lines) - get-settings, set-setting

zones.go (144 lines) - types, list, icons, findZone helper
zones_manage.go (110 lines) - rename, set-icon, delete

Added shared helper functions (findDevice, findZone, parseValue) to
reduce code duplication across commands.
